feat: create seperate utilities for resource loading
