--- src/utilities/utility-helper.js
@@ -1,242 +1,0 @@
-export const BaseResources = new Map([
-    ["sfx-fast", { type: "sounds", name: "sfx-fast", ext: "wav" }],
-    ["sfx-slow", { type: "sounds", name: "sfx-slow", ext: "wav" }],
-    ["sfx-jump", { type: "sounds", name: "sfx-jump", ext: "wav" }],
-    ["sfx-lose", { type: "sounds", name: "sfx-lose", ext: "wav" }],
-    ["sprite-player", { type: "spritesheets", name: "sprite-player", ext: "png" }],
-    ["sprite-enemy-spike", { type: "spritesheets", name: "sprite-enemy-spike", ext: "png" }],
-    ["sprite-hud", { type: "spritesheets", name: "sprite-hud", ext: "png" }],
-    ["sprite-items", { type: "spritesheets", name: "sprite-items", ext: "png" }],
-    ["tileset-platform", { type: "tilesets", name: "tileset-platform", ext: "png" }],
-    ["background-hills", { type: "images", name: "background-hills", ext: "png" }]
-]);
-
-export const ItemTypes = {
-    Coffee: 'COFFEE',
-    Camomile: 'CAMOMILE',
-    Chocolate: 'CHOCOLATE'
-};
-
-export const SpeedTypes = {
-    Normal: 60,
-    Slow: 30,
-    Fast: 90,
-    Jump: 240,
-    DoubleJump: 280
-};
-
-export const LoseTileTypes = {
-    Cliff: 1,
-    Spikes: 2
-};
-
-export const TileSettings = {
-    TileSize: 8,
-    TileRows: 11,
-    TileGroundLevel: 3,
-    TileChunkDefaultSize: 88,
-    TileChunkDefaultActive: 4,
-};
-
-export class Utils {
-    static isValueEmpty = (resourceValue) => {
-        if (resourceValue === 0 || resourceValue === undefined || resourceValue === "") {
-            return true;
-        }
-        return false;
-    }
-
-    static loadResources = (scene) => {
-        for (const [key, value] of BaseResources) {
-            let resourcePath = `assets/${value.type}/${value.name}.${value.ext}`;
-            switch (value.type) {
-                case "sounds":
-                    scene.load.audio(value.name, resourcePath);
-                    break;
-                case "images":
-                case "tilesets":
-                    scene.load.image(value.name, resourcePath);
-                    break;
-                case "spritesheets":
-                    scene.load.spritesheet(value.name, resourcePath, { frameWidth: 8, frameHeight: 8, margin: 0, spacing: 0 });
-                    break;
-                default:
-                    break;
-            }
-        }
-    }
-
-    static createAnimations = (scene) => {
-        const anims = scene.anims;
-        if (anims.anims.size === 0) {
-            anims.create({
-                key: "player-idle",
-                frames: anims.generateFrameNumbers("sprite-player", { start: 0, end: 3 }),
-                frameRate: 3,
-                repeat: -1,
-            });
-            anims.create({
-                key: "player-run",
-                frames: anims.generateFrameNumbers("sprite-player", { start: 8, end: 15 }),
-                frameRate: 12,
-                repeat: -1,
-            });
-            anims.create({
-                key: "player-destroy",
-                frames: anims.generateFrameNumbers("sprite-player", { start: 16, end: 23 }),
-                frameRate: 8,
-                repeat: 0,
-            });
-            anims.create({
-                key: "enemy-idle",
-                frames: anims.generateFrameNumbers("sprite-enemy-spike", { start: 0, end: 3 }),
-                frameRate: 4,
-                repeat: -1,
-            });
-            anims.create({
-                key: "enemy-walk",
-                frames: anims.generateFrameNumbers("sprite-enemy-spike", { start: 4, end: 7 }),
-                frameRate: 4,
-                repeat: -1,
-            });
-            anims.create({
-                key: "coffee",
-                frames: anims.generateFrameNumbers("sprite-items", { start: 0, end: 3 }),
-                frameRate: 4,
-                repeat: -1,
-            });
-            anims.create({
-                key: "camomile",
-                frames: anims.generateFrameNumbers("sprite-items", { start: 4, end: 7 }),
-                frameRate: 4,
-                repeat: -1,
-            });
-            anims.create({
-                key: "chocolate",
-                frames: anims.generateFrameNumbers("sprite-items", { start: 8, end: 11 }),
-                frameRate: 4,
-                repeat: -1,
-            });
-        }
-    }
-
-    static createBackgrounds = (scene, count, texture, scrollFactor) => {
-        let trackingXValue = 0;
-        let images = [];
-        for (let i = 0; i < count; ++i) {
-            const currentImage = scene.add.image(trackingXValue, 0, texture)
-                .setOrigin(0, 0)
-                .setScrollFactor(scrollFactor);
-            trackingXValue += currentImage.width;
-            images.push(currentImage);
-        }
-        return images;
-    }
-
-    static createSounds = (scene) => {
-        for (const [key, value] of BaseResources) {
-            if (value.type === "sounds") {
-                if (value.name.includes("jump")) {
-                    scene.jumpSound = scene.sound.add(value.name);
-                }
-                if (value.name.includes("lose")) {
-                    scene.loseSound = scene.sound.add(value.name);
-                }
-                if (value.name.includes("fast")) {
-                    scene.fastSound = scene.sound.add(value.name);
-                }
-                if (value.name.includes("slow")) {
-                    scene.slowSound = scene.sound.add(value.name);
-                }
-            }
-        }
-    }
-
-    static runLoseSequenceDynamic(scene, currentStage, time, byFall) {
-        scene.time.delayedCall(time, () => {
-            currentStage++;
-            if (byFall === true) {
-                if (currentStage === 1) {
-                    if (scene.loseSequenceSound === false) {
-                        scene.loseSound.play();
-                        scene.loseSequenceSound = true;
-                    }
-                    scene.cameras.main.stopFollow();
-                    this.runLoseSequenceDynamic(scene, currentStage, time * 400, byFall);
-                } else {
-                    scene.scene.restart();
-                }
-            } else {
-                switch (currentStage) {
-                    case 1:
-                        if (scene.player.sprite.anims) {
-                            scene.player.sprite.anims.play("player-destroy", true);
-                        }
-                        this.runLoseSequenceDynamic(scene, currentStage, time, byFall);
-                        break;
-                    case 2:
-                        if (scene.loseSequenceSound === false) {
-                            scene.loseSound.play();
-                            scene.loseSequenceSound = true;
-                        }
-                        scene.cameras.main.stopFollow();
-                        this.runLoseSequenceDynamic(scene, currentStage, time * 200, byFall);
-                        break;
-                    default:
-                        scene.scene.restart();
-                }
-            }
-        }, [], this);
-    }
-
-    static createItemSequence = (playerReference, itemSprite) => {
-        const scene = playerReference.scene;
-        const item = scene.items.find(i => i.sprite === itemSprite);
-        if (item && item.activated === false) {
-            item.activated = true;
-            itemSprite.setVisible(false);
-            scene.player.fastSequenceActive = false;
-            scene.player.slowSequenceActive = false;
-            if (item.type === ItemTypes.Coffee) {
-                scene.player.fastSequenceActive = true;
-                scene.fastSound.play();
-            }
-            if (item.type === ItemTypes.Camomile) {
-                scene.player.slowSequenceActive = true;
-                scene.slowSound.play();
-            }
-            if (item.type === ItemTypes.Chocolate) {
-                scene.fastSound.play();
-                scene.hudJumpBarCounter++;
-                if (scene.hudJumpBarCounter > 3) {
-                    scene.hudJumpBarCounter = 3;
-                }
-                this.setHudBar(scene);
-            }
-            scene.time.delayedCall(2000, () => {
-                scene.player.fastSequenceActive = false;
-                scene.player.slowSequenceActive = false;
-            });
-        }
-    }
-
-    static setHudBar = (sceneReference) => {
-	switch (sceneReference.hudJumpBarCounter) {
-	    case 0:
-		sceneReference.hudBar.setFrame(16);
-		break;
-	    case 1:
-		sceneReference.hudBar.setFrame(17);
-		break;
-	    case 2:
-		sceneReference.hudBar.setFrame(18);
-		break;
-	    case 3:
-		sceneReference.hudBar.setFrame(19);
-		break;
-	    default:
-		break;
-	}
-    }
-}
-
